\name{pllr}
\alias{pllr}
\title{Penalized Local Linear Regression}
\description{
Data sharpened local linear regression with roughness penalty with
automatically selected bandwidth and tuning parameter.
}
\usage{
pllr(x, y)
}
\arguments{
\item{x}{numeric vector of predictor observations}
\item{y}{numeric vector of observed responses}
}
\value{
a list consisting of the x and y coordinates of the estimated
regression function.
}
\author{W.J. Braun}
\keyword{models}


